+++ docs/changes.rst
@@ -5,6 +5,11 @@ Next version
 ------------
 
 * Ensured that the handle stays within bounds when resizing the window.
+* Disabled ``HistoryPanel`` when ``RENDER_PANELS`` is ``True``
+  or if ``RENDER_PANELS`` is ``None`` and the WSGI container is
+  running with multiple processes.
+* Fixed ``RENDER_PANELS`` functionality so that when ``True`` panels are
+  rendered during the request and not loaded asynchronously.
 
 
 3.2.1 (2021-04-14)